+++ types/aws-lambda/trigger/sns.d.ts
@@ -1,0 +1,38 @@
+import { Handler } from "../hander";
+
+export type SNSHandler = Handler<SNSEvent, void>;
+
+// SNS "event"
+export interface SNSMessageAttribute {
+    Type: string;
+    Value: string;
+}
+
+export interface SNSMessageAttributes {
+    [name: string]: SNSMessageAttribute;
+}
+
+export interface SNSMessage {
+    SignatureVersion: string;
+    Timestamp: string;
+    Signature: string;
+    SigningCertUrl: string;
+    MessageId: string;
+    Message: string;
+    MessageAttributes: SNSMessageAttributes;
+    Type: string;
+    UnsubscribeUrl: string;
+    TopicArn: string;
+    Subject: string;
+}
+
+export interface SNSEventRecord {
+    EventVersion: string;
+    EventSubscriptionArn: string;
+    EventSource: string;
+    Sns: SNSMessage;
+}
+
+export interface SNSEvent {
+    Records: SNSEventRecord[];
+}
